+++ src/main.rs
@@ -36,14 +36,14 @@ fn main() -> ExitCode {
         Kay::Check( mut check ) => match check.execute() {
             Ok( _ ) => ExitCode::SUCCESS,
             Err( err ) => {
-                eprintln!( "{}", err );
+                eprint!( "{}", err );
                 ExitCode::FAILURE
             },
         },
         Kay::Compile( mut compile ) => match compile.execute() {
             Ok( _ ) => ExitCode::SUCCESS,
             Err( err ) => {
-                eprintln!( "{}", err );
+                eprint!( "{}", err );
                 ExitCode::FAILURE
             },
         },
@@ -70,42 +70,16 @@ mod tests {
                     }.into();
 
                     match kay {
-                        // Kay::Help( help ) => {
-                        //     help.execute();
-                        //     ExitCode::SUCCESS
-                        // },
-                        // Kay::Version( version ) => {
-                        //     version.execute();
-                        //     ExitCode::SUCCESS
-                        // },
                         Kay::Check( mut check ) => match check.execute() {
                             Ok( _ ) => {
                             },
                             Err( err ) => {
-                                eprintln!( "{}", err );
+                                eprint!( "{}", err );
                                 return ExitCode::FAILURE;
                             },
                         },
                         _ => unreachable!(),
-                        // Kay::Compile( mut compile ) => match compile.execute() {
-                        //     Ok( _ ) => {
-                        //         compile.logger.done();
-                        //         ExitCode::SUCCESS
-                        //     },
-                        //     Err( err ) => {
-                        //         eprintln!( "{}", err );
-                        //         ExitCode::FAILURE
-                        //     },
-                        // },
                     }
-
-                    // match kay.execute() {
-                    //     Ok( () ) => {},
-                    //     Err( err ) => {
-                    //         eprintln!( "{}", err );
-                    //         return ExitCode::FAILURE;
-                    //     },
-                    // }
                 }
             }
         }